Make getRecommendation request nonblocking
Previously, getRecommendations request was a blocking call and
until server responds, devportal is showing the loading sign. Since
server is always in a separate network, thread blocking call to get
recommendations can be an issue as there can be network latencies.
With new changes, we have introduced a cache. Recommended API
list is obtained from the cache and this cache will be updated in a
non blocking thread, every time a new event published.

Signed-off-by: Lakshitha Gunasekara <lakshitha.gunasekara@gmail.com>