Improved clicked_api publishing payload
Previously when an user clicks an API on devportal, the name
of the API is published to the recommendation server. This will
create an error if there is another API with the same name in a
different tenant or a company. New changes will publish the
API ID, which is unique for each API.

Signed-off-by: Lakshitha Gunasekara <lakshitha.gunasekara@gmail.com>